cryptoengine/services/backtester/requirements.txt
pandas>=2.2.0,<3.0
numpy>=1.26.0,<2.0
matplotlib>=3.7.0,<4.0
asyncpg>=0.29.0,<1.0
pydantic>=2.0,<3.0
structlog>=24.0,<25.0
pyyaml>=6.0,<7.0
ccxt>=4.3.0,<5.0
psycopg2-binary>=2.9.0,<3.0
redis[hiredis]>=5.0,<6
hmmlearn==0.3.2
scipy>=1.12.0,<2.0
scikit-learn>=1.4.0,<2.0
beautifulsoup4>=4.12.0,<5.0
requests>=2.31.0,<3.0
aiohttp>=3.9.0,<4.0
statsmodels>=0.14.0,<1.0
xgboost>=2.0.0,<3.0
shap>=0.45.0,<1.0
pyarrow>=15.0.0,<16.0
